pyproject.toml
[project]
name = "major-project"
version = "1.0.0"
description = "Major-Project AI Pentester Agent - Autonomous security testing with human oversight for continuous attack surface discovery and vulnerability assessment"
requires-python = ">=3.12"
dependencies = [
    "agentex-sdk>=0.6.5",
    "temporalio>=1.18.2",
    "openai-agents>=0.4.2",  # OpenAI Agents SDK for AI-powered reasoning
    "litellm>=1.30.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "termcolor>=2.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "redis>=5.0.0",
    "motor>=3.3.0",  # MongoDB async driver for attack surface persistence
    "pyyaml>=6.0.0",  # YAML parsing for OpenAPI specs
    "aiosmtplib>=3.0.0",  # Async SMTP for email alerts
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["project"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
